pyproject.toml
[project]
name = "model-explorer-onnx"
version = "0.0.1"

dependencies = [
  "ai-edge-model-explorer"
]
